+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -9,6 +9,7 @@
 #### Fixed
 - Incorrect `SYSIN_MEMBER` variable value
 - Missing copybook reported in inactive code
+- Sequence symbol redefinition reported in the lookahead mode
 
 ## [1.20.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.19.0...1.20.0) (2025-11-12)